refactor: 컨트롤러의 패키지 구조 변경
팀원들의 룰에 따라 패키지 구조 변경
