+++ src/main/java/upbrella/be/user/controller/UserController.java
@@ -1,4 +1,4 @@
-package upbrella.be.user;
+package upbrella.be.user.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
